new file:   image/.gitignore
Introducing a .gitignore to the image path.
The motivation is to be able to introduce symbolic links to directories containing Dockerfiles.
They would be accessible by the Makefile, but not noticed by Git.
This is helpful for extending the Dockerfiles without having to manage Docker from multiple paths.
Otherwise, you're limited to 'make build n=base|node'.