modified:   image/node/Dockerfile
Modified the Node Dockerfile to expose port 8080 instead of 80 and 443.
The motivation is that port numbers in range 0 to 1023 are 'well-known' ports. They are used by system processes that provide widely used types of network services, and because of this, a process must execute with superuser priviliges to be able to bind a network socket to an IP address on one of these 'well-known' ports.

8080 is an officially registered, alternate port with IANA for TCP and HTTP as a non-root user.