modified:   git-shell-commands/git
Modified the Git shell command 'git.'
The Git shell commands only get invoked by the git user.
The Git template directory was NOT being used by the git user before because it was set as a global that I believe was only accessible by the root user.
The Git template directory is now being passed in as an option in the Git init command invocation.
Example: git --bare init --template='/home/git/.git-templates'