new file:   bootstrap.sh
Bootstrap.sh is a shell script being introduced to bootstrap the Paasta runtime environment on any Docker-capable host.
It attempts to download the Docker installation script using either wget or curl in any Posix-capable shell environment, and then execute it as root.
The installation script should take care of platform introspection for things such as kernel version, and 64 bit detection, and OS version.
The recommended OS is Ubuntu 12.04, but or *nix distributions are allowed.