deleted:    image/base/bin/entry
Deleted the entry point shell script because it is no loner necessary to pass in an SSH public key and start the SSHD from shell.
Codebox is now the default entry point to any container. SSH keys may be managed from that interface.
SSH is not even necessary unless requested.
SupervisorD is being used to auto start and restart the SSHD now.