new file:   push.sh
Added a shell script called push.sh to allow for pushing the Paasta runtime to any remote host.
It accepts any path, and any host/path, and attempts to connect to it in an encrypted tunnel using SSH.
It tries to connect using Rsync, and falling back to Scp. If neither are available, it errors.