modified:   image/base/supervisor/supervisord.conf
Modified the supervisord.conf file to remove the default username and password for each Codebox instance.
The intent is to put in HTTP Basic Authentication in the reverse proxy layer to authenticate users to their Codebox.
This will make it easier to administer users for each Codebox external to the Codebox's respective Docker container.